Refactor db layer, fix bugs

+++ api/src/main/java/ru/fitness/dao/IProg.java
@@ -3,7 +3,7 @@ package ru.fitness.dao;
 import java.util.Set;
 
 public interface IProg {
-    Integer getPrevProgId();
+    Long getPrevProgId();
 
     boolean isPrevious();
